Readdressed all Dev-C issues.  Should compile correctly in both MSVC and Dev-C (though Dev-C will give you many warnings.  Just the way of things).
Started a Bot Debugging window.  Wolf, could you check out if it's put together more or less correctly?

+++ GFX/Primitives.cpp
@@ -25,7 +25,7 @@ void CreateCircle(Vector4 c, float radius, long divisions)
     for(int i =0;i<divisions;i++)
 	{
       angle = float(i*2*PI/divisions);
-      glVertex2f(c.x()+cos(angle)*radius,c.y()+sin(angle)*radius);
+      glVertex2f(c.x()+cosf(angle)*radius,c.y()+sinf(angle)*radius);
     }
 	glEnd();
 }
